requirements.txt
absl-py==1.4.0
appdirs==1.4.4
attributee==0.1.8
attrs==23.1.0
bidict==0.22.1
brotlipy==0.7.0
cachetools==5.3.1
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
chardet @ file:///tmp/build/80754af9/chardet_1607706768982/work
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography @ file:///croot/cryptography_1677533068310/work
cycler==0.11.0
Cython==3.0.0
dataclasses==0.6
docker-pycreds==0.4.0
dominate==2.8.0
easydict==1.10
fonttools==4.38.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==0.4.6
grpcio==1.56.2
idna @ file:///croot/idna_1666125576474/work
importlib-metadata==6.7.0
importlib-resources==5.12.0
info-nce-pytorch==0.1.4
install==1.3.5
jpeg4py==0.1.4
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.17.3
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
llvmlite==0.39.1
lmdb==1.4.1
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.5.3
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
networkx==2.6.3
numba==0.56.4
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1653915516269/work
oauthlib==3.2.2
opencv-python==4.8.0.74
ordered-set==4.1.0
packaging==23.1
pandas==1.3.5
pathtools==0.1.2
phx-class-registry==4.0.6
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
protobuf==3.20.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.6
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
PyLaTeX==1.4.1
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing==3.1.0
pyrsistent==0.19.3
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.7.3
sentry-sdk==1.30.0
setproctitle==1.3.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
tb-nightly==2.12.0a20230113
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorly==0.8.1
timm==0.5.4
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tornado==6.2
tqdm @ file:///opt/conda/conda-bld/tqdm_1664392687731/work
typing_extensions @ file:///tmp/abs_ben9emwtky/croots/recipe/typing_extensions_1659638822008/work
urllib3==1.26.16
visdom==0.2.4
vot-toolkit==0.5.3
vot-trax==3.0.2
wandb==0.15.9
websocket-client==1.6.1
Werkzeug==2.2.3
zipp==3.15.0
